update minishell parsing

+++ circle03/minishell/includes/utils/handle_files.h
@@ -2,6 +2,6 @@
 
 # define HANDLE_FILE_H
 
-int    handle_redirection(char **tokens, size_t i, t_cmd *node);
+int    handle_redirection(char **tokens, int i, t_cmd *node);
 
 #endif
